requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coverage==6.3.2
cryptography==36.0.1
Django==4.0.2
django-countries==7.2.1
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-mptt==0.13.4
factory-boy==3.2.1
Faker==13.2.0
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
numpy==1.22.2
packaging==21.3
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.65.0
stripe3==0.0.0
tomli==2.0.1
tzdata==2021.5
urllib3==1.26.8
